+++ listen-kit/src/hype/mod.rs
@@ -1,19 +1,17 @@
 use anyhow::Result;
-use hyperliquid_rust_sdk::{
-    BaseUrl, ClientOrder, ClientOrderRequest, ClientTrigger, ExchangeClient,
-    InfoClient,
-};
+use ethers::types::H160;
 use rig::{agent::AgentBuilder, streaming::StreamingCompletionModel};
-use rig_tool_macro::tool;
 
 use crate::agent::Features;
-use crate::common::{spawn_with_signer, OpenRouterAgent};
+use crate::common::OpenRouterAgent;
+use crate::hype::info::*;
+use crate::hype::orders::*;
 #[cfg(feature = "hype")]
-use crate::signer::AsHypeSigner;
-use crate::signer::SignerContext;
 use crate::{
     agent::model_to_versioned_model, common::openrouter_agent_builder,
 };
+pub mod info;
+pub mod orders;
 
 const PREAMBLE: &str = "You are a Hyperliquid assistant. Hyperliquid is a
 high-performance decentralized derivatives exchange that processes billions of
@@ -22,90 +20,15 @@ through specialized tools and aim to help users understand and interact with the
 platform effectively. You can provide information about prices, order books, and
 market conditions to help users make informed decisions, as well as execute orders on their behalf.";
 
-#[tool(description = "
-Gets the complete orderbook snapshot for a given coin. Example response:
-{
-  \"coin\": \"ETH\",
-  \"levels\": [
-    [
-      {\"n\": 1, \"px\": \"2545.4\", \"sz\": \"11.7811\"}, // 1 order at 2545.4, size 11.7811 ETH
-      {\"n\": 12, \"px\": \"2545.0\", \"sz\": \"136.8789\"}, // 12 orders at 2545.0, size 136.8789 ETH
-      {\"n\": 17, \"px\": \"2544.9\", \"sz\": \"144.4251\"}, // 17 orders at 2544.9, size 144.4251 ETH
-      // ... more orders deeper on the bid (buy) side, skipped for brevity
-    ],
-    [
-      {\"n\": 1, \"px\": \"2545.5\", \"sz\": \"0.0061\"}, // 1 order at 2545.5, size 0.0061 ETH
-      {\"n\": 10, \"px\": \"2545.6\", \"sz\": \"40.0728\"}, // 10 orders at 2545.6, size 40.0728 ETH
-      {\"n\": 6, \"px\": \"2545.7\", \"sz\": \"102.1028\"}, // 6 orders at 2545.7, size 102.1028 ETH
-      // ... more orders deeper on the ask (sell) side, skipped for brevity
-    ]
-  ],
-  \"time\": 1748279333332
-}")]
-pub async fn get_l2_snapshot(coin: String) -> Result<serde_json::Value> {
-    // thread-local this, possibly onto the signer
-    let client = InfoClient::new(None, Some(BaseUrl::Mainnet)).await?;
-    let info = client.l2_snapshot(coin).await?;
-    Ok(serde_json::to_value(info)?)
-}
-
-#[tool(description = "
-Send a market order to the exchange.
-{
-  \"coin\": \"ETH\",
-  \"side\": \"buy\",
-  \"size\": \"0.01\",
-}
-
-{
-  \"coin\": \"ETH\",
-  \"side\": \"sell\",
-  \"size\": \"0.01\",
-}
-")]
-pub async fn send_market_order(
-    coin: String,
-    side: String,
-    size: String,
-) -> Result<serde_json::Value> {
-    let signer = SignerContext::current().await;
-    spawn_with_signer(signer.clone(), move || async move {
-        let client = ExchangeClient::new(
-            None,
-            signer.as_hype_signer(),
-            Some(BaseUrl::Mainnet),
-            None,
-            None,
-        )
-        .await?;
-        let info = client
-            .order(
-                ClientOrderRequest {
-                    asset: coin,
-                    is_buy: side == "buy",
-                    reduce_only: false,
-                    limit_px: 0.,
-                    sz: size.parse::<f64>()?,
-                    cloid: None,
-                    order_type: ClientOrder::Trigger(ClientTrigger {
-                        is_market: true,
-                        trigger_px: 0.,
-                        tpsl: "".to_string(),
-                    }),
-                },
-                Some(&*signer.as_hype_signer()),
-            )
-            .await?;
-        Ok(serde_json::to_value(info)?)
-    })
-    .await
-    .await?
-}
-
 pub fn equip_with_hype_tools<M: StreamingCompletionModel>(
     agent: AgentBuilder<M>,
 ) -> AgentBuilder<M> {
-    agent.tool(GetL2Snapshot).tool(SendMarketOrder)
+    agent
+        .tool(GetL2Snapshot)
+        .tool(SendMarketOrder)
+        .tool(GetOpenOrders)
+        .tool(DepositUsdc)
+        .tool(GetBalanceOverview)
 }
 
 pub fn create_hype_agent_openrouter(
@@ -118,4 +41,10 @@ pub fn create_hype_agent_openrouter(
         .preamble(PREAMBLE)
         .build();
     agent
+}
+
+pub fn parse_evm_address(address: Option<String>) -> Result<H160> {
+    Ok(address
+        .ok_or(anyhow::anyhow!("No EVM address found"))?
+        .parse::<H160>()?)
 }
